Palautettava versio

+++ src/entities/account.py
@@ -5,8 +5,8 @@ class Account:
         name: name of the account as a string
         password: Password of the account as a string
         directory: Directory of the account, where all output files will be saved, as a string
-        primer_length: Length of the sequencing primer. Default value 20
-        primer_gc_content: GC content of the sequencing primer. Default value of 0.5
+        primer_length: Length of the sequencing primer as integer. Default value 20
+        primer_gc_content: GC content of the sequencing primer as float. Default value of 0.5
     """
 
     def __init__(self, name, password, directory, primer_length=20, primer_gc_content=0.5):
@@ -16,8 +16,8 @@ class Account:
             name: name of the account as a string
             password: Password of the account as a string
             directory: Directory of the account, where all output files will be saved, as a string
-            primer_length: Length of the sequencing primer. Default value 20
-            primer_gc_content: GC content of the sequencing primer. Default value of 0.5
+            primer_length: Length of the sequencing primer as integer. Default value 20
+            primer_gc_content: GC content of the sequencing primer as float. Default value of 0.5
         """
 
         self.name = name
